feat(resources): add empty index page; remove old resources pages
